move metrics interface back
